+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Modified `OutputPin` behavior for active-low pins to match `InputPin` behavior.
 - Set default features to build both sysfs and cdev pin types.
 - Removed `Pin` export, use `CdevPin` or `SysfsPin`.
+- Increased the Minimum Supported Rust Version to `1.46.0` due to an update of `bitflags`.
 - Adapted to `embedded-hal` `1.0.0-alpha.5` release.
 - Updated `nb` to version `1`.