+++ CHANGELOG.md
@@ -17,8 +17,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Set default features to build both sysfs and cdev pin types.
 - Removed `Pin` export, use `CdevPin` or `SysfsPin`.
-- Updated `embedded-hal` to version 0.2.6.
+- Updated `embedded-hal` to version `0.2.6`.
 - Updated `nb` to version `0.1.3`.
+- Updated `gpio-cdev` to version `0.5`.
+- Updated `i2cdev` to version `0.5`.
+- Updated `spidev` to version `0.5`.
+- Updated `sysfs-gpio` to version `0.6`.
+- Updated `cast` to version `0.3`.
 
 ### Fixed